# backend/qcloudy-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QCloudy API",
    "summary": "Player profile and transformed SkyBlock market data for QCloudy Addition.",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "Operations"
        ],
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/ready": {
      "get": {
        "tags": [
          "Operations"
        ],
        "summary": "Ready",
        "operationId": "ready_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Ready Ready Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/pv/{target}": {
      "get": {
        "tags": [
          "Player Viewer"
        ],
        "summary": "Player Viewer",
        "operationId": "player_viewer_v1_pv__target__get",
        "parameters": [
          {
            "name": "target",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 36,
              "title": "Target"
            }
          },
          {
            "name": "profileId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 32,
                  "maxLength": 36
                },
                {
                  "type": "null"
                }
              ],
              "title": "Profileid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Player Viewer V1 Pv  Target  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/pv/{player_uuid}/{profile_id}/museum": {
      "get": {
        "tags": [
          "Player Viewer"
        ],
        "summary": "Museum",
        "operationId": "museum_v1_pv__player_uuid___profile_id__museum_get",
        "parameters": [
          {
            "name": "player_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 32,
              "maxLength": 36,
              "title": "Player Uuid"
            }
          },
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 32,
              "maxLength": 36,
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Museum V1 Pv  Player Uuid   Profile Id  Museum Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/pv/{player_uuid}/{profile_id}/garden": {
      "get": {
        "tags": [
          "Player Viewer"
        ],
        "summary": "Garden",
        "operationId": "garden_v1_pv__player_uuid___profile_id__garden_get",
        "parameters": [
          {
            "name": "player_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 32,
              "maxLength": 36,
              "title": "Player Uuid"
            }
          },
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 32,
              "maxLength": 36,
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Garden V1 Pv  Player Uuid   Profile Id  Garden Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/market/prices": {
      "post": {
        "tags": [
          "Market"
        ],
        "summary": "Prices",
        "operationId": "prices_v1_market_prices_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchPriceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Prices V1 Market Prices Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/market/bazaar/shards": {
      "get": {
        "tags": [
          "Market"
        ],
        "summary": "Shard Prices",
        "operationId": "shard_prices_v1_market_bazaar_shards_get",
        "parameters": [
          {
            "name": "side",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "instant_buy",
                "instant_sell"
              ],
              "type": "string",
              "default": "instant_buy",
              "title": "Side"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Shard Prices V1 Market Bazaar Shards Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/market/status": {
      "get": {
        "tags": [
          "Market"
        ],
        "summary": "Market Status",
        "operationId": "market_status_v1_market_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Market Status V1 Market Status Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BatchPriceRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PriceItemRequest"
            },
            "type": "array",
            "maxItems": 256,
            "minItems": 1,
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "BatchPriceRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PriceItemRequest": {
        "properties": {
          "itemId": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Itemid"
          },
          "variantKey": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Variantkey"
          }
        },
        "type": "object",
        "required": [
          "itemId"
        ],
        "title": "PriceItemRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}